[YITEAM-74] Follow 관련 API User -> Follow 수정
