+++ src/main/java/com/yapp/ios1/controller/FollowController.java
@@ -2,16 +2,25 @@ package com.yapp.ios1.controller;
 
 import com.yapp.ios1.common.ResponseMessage;
 import com.yapp.ios1.dto.ResponseDto;
+import com.yapp.ios1.dto.user.result.FriendDto;
 import com.yapp.ios1.service.FollowService;
 import com.yapp.ios1.utils.auth.Auth;
 import com.yapp.ios1.utils.auth.UserContext;
 import io.swagger.annotations.ApiOperation;
+import io.swagger.annotations.ApiResponse;
+import io.swagger.annotations.ApiResponses;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
+
+import static com.yapp.ios1.common.ResponseMessage.GET_FRIEND_LIST;
+import static com.yapp.ios1.common.ResponseMessage.NO_FRIEND_LIST;
 
 /**
  * created by jg 2021/05/21
@@ -33,5 +42,23 @@ public class FollowController {
         followService.followRequest(myUserId, friendId);
         return ResponseEntity.ok()
                 .body(ResponseDto.of(HttpStatus.CREATED, ResponseMessage.FRIEND_REQUEST));
+    }
+
+    @ApiOperation(
+            value = "친구 리스트"
+    )
+    @ApiResponses({
+            @ApiResponse(code = 200, message = "친구 목록 존재하는 경우"),
+            @ApiResponse(code = 404, message = "친구 목록 존재하지 않는 경우"),
+    })
+    @GetMapping("/users/{userId}/friends")
+    public ResponseEntity<ResponseDto> getFriendList(@PathVariable Long userId) {
+        List<FriendDto> friendList = followService.getFriendList(userId);
+        if (friendList.size() == 0) {
+            return ResponseEntity.ok()
+                    .body(ResponseDto.of(HttpStatus.NOT_FOUND, NO_FRIEND_LIST));
+        }
+        return ResponseEntity.ok()
+                .body(ResponseDto.of(HttpStatus.OK, GET_FRIEND_LIST, friendList));
     }
 }
